Kill libinit
I've never liked this, I prefer properties to be in a easily readable file once system is installed.
Plus this caused some weird denials to me in the past and it requires additional "hackery" to system/core.
This is going to be replaced by a simple script run at boot time.